Refactor hitevents and support heal events
Mainly this is a refactoring job, with the added bonus of getting
healevents in the damage account panel as well. Heal and hurt events are
very similar, so the code paths are shared.

The former hook was on `CDamageAccountPanel::DisplayDamageFeedback` but
since it was needed to hook `CDamageAccountPanel::FireGameEvent` anyway
to support heal events(a `localPlayer->m_lifeState = LIFE_ALIVE` check)
the module no longer has its own `FireGameEvent` anymore, instead moving
the required functionality directly to the hook.