Fix disappearing particle effects e.g. heal beams
On SourceTV medic beams disappear shortly after the medic starts
healing, and unusual weapon effects have the same problem.

The problem is caused by `C_TFWeaponBase::PostDataUpdate` not correctly
comparing the model index when receiving it from the server. Because
of how weapon models works, special handling is needed when in first
person view because the same backing model is used for world and view
models now.

`C_BasePlayer::ShouldDrawLocalPlayer` is the check that determines which
model we are to use, and because it doesn't handle SourceTV is picks the
wrong one, with model overrides and disappearing particle collections as
a result.

Because `ShouldDrawLocalPlayer` doesn't have the context needed to check
if a given weapon belongs to the currently viewed player, so we can't
simply hook it and override its results. Luckily, it turns out that we
don't need to override it in general, just when it's called from
`PostDataUpdate`. So we just hook both and keep the weapon owner as
state(managed by a variable pusher), and this allows us to check if the
given model belongs to the local player.

The fix is behind the cvar `ce_graphics_fix_viewmodel_particles` which
defaults to `1`.

I decided to put this in the `Graphics` module because it seems like a
graphical fix, so most people will likely look for it here. This does
introduce a soft dependency on `CameraState`, as we use the
`GetLocalObserverMode` and `GetLocalObserverMode` helpers to figure out
the correct states, but should `CameraState` break the fix can simply be
disabled with the `cvar` such that CE continues to function.