Allow use of client scheme named colors (#93)
* Allow use of client scheme named colors

Here we update the generic `ColorFromString` function to parse names
from the client scheme.

Since the `IScheme` interface doesn't allow us to query for the
existence of a given color, we use a small hack here, by querying the
scheme twice instead.

Fixes #1.

* Small optimization in `ColorFromString`

When `ColorFromString` performs a scheme color lookup, we always did two
looksup so we could figure out if the color was present or not.

We can actually do a small short-circuit evaluation here, and only do
the second lookup if the color is the color we specified as the default
color in the first lookup. Since this default color is picked as
(254,1,254) it's quite unlikely someone would ever pick it as a real
color.