Support for Enthusiast's Timepiece & Quackenbirdt
I'm not too well-versed in C++, and there's probably a classier way of declaring all of the item IDs into one chargebar declaration, but this gets the job done. This implementation also allows the reskins' names to be adjusted depending on the item, though by default they all show up as the stock invis watch since they're just reskins.