Fix UberCharge status remaining in HUD
Sometimes the server will correctly clear `TF_COND_INVULNERABLE` but not
`TF_COND_INVULNERABLE_WEARINGOFF`, which causes the status icon to stick
around until the wearing off effect is cleared, typically after the next
time the player receives an uber.

As spectators we are only really interested in the fact that the player
can die or not, so by simply not showing the uber status when only the
`_WEARINGOFF` condition is set we can

This might also fix
https://github.com/PazerOP/CastingEssentials/issues/78 but I'm not sure
based on the issue text alone.