Interfaces+HitEvents+SrcTVPlus: Remove `C_BasePlayer::GetLocalPlayer`
There is no reason to signature scan for this -- the Engine can easily
give us this, and we do not hook the function itself.

The usage of this was flawed in the end, as the `C_BasePlayer*` was
cached for some bizarre reason, causing it to become invalid as soon as
it was deleted (full network update, changing demo/STV server).

This was causing both hit events and SrcTV+ detection to break after the
first demo/server.