SrcTVPlus: Simplify comparison between local player and entity in proxy
This is an easier and more correct comparison rather than comparing
player entity pointers -- just compare the entity indices.

It is more correct because it previously assumed the local player had
already been transmitted (non-null). This is normally the case, because
the STV player index is almost always `1`.

However, nothing requires the STV player to be `1`. In cases where, for
example, it is `2`, the player index `1` will be transmitted, and
attempt to be compared to the local player entity pointer. This will
result in a crash due to our boxing around the local player and it's
base entity, which hasn't been transmitted yet.